fix(ingress): keep standard profile within line limits
