+++ src/channels/message/ingress-monitor.ts
@@ -52,7 +52,7 @@ export function createChannelIngressError(
 }
 
 /** Stable identity and serialization lane extracted before durable admission. */
-type ChannelIngressMonitorFacts = { eventId: string; laneKey: string };
+export type ChannelIngressMonitorFacts = { eventId: string; laneKey: string };
 
 /** Versioned body presented to a channel's persisted-payload encoder. */
 type ChannelIngressPayloadEnvelope<TBody> = { version: number; body: TBody };
@@ -84,7 +84,7 @@ type ChannelIngressMonitorInspectionContext =
 
 type ChannelIngressMonitorClaimErrorKind = "invalid-version" | "identity-mismatch";
 
-type ChannelIngressMonitorPayloadCodec<TRaw, TBody, TStoredPayload, TMetadata> = {
+export type ChannelIngressMonitorPayloadCodec<TRaw, TBody, TStoredPayload, TMetadata> = {
   version: number;
   serialize: (
     raw: TRaw,
@@ -129,12 +129,12 @@ export const CHANNEL_INGRESS_RETENTION_DEFAULTS = Object.freeze({
   failedMaxEntries: 20_000,
 } satisfies ChannelIngressMonitorRetention);
 
-type ChannelIngressMonitorDrainOptions<TStoredPayload, TMetadata> = Omit<
+export type ChannelIngressMonitorDrainOptions<TStoredPayload, TMetadata> = Omit<
   CreateChannelIngressDrainOptions<TStoredPayload, TMetadata>,
   "queue" | "dispatchClaimedEvent" | "abortSignal" | "now" | "ownerId" | "claimLeaseMs"
 >;
 
-type CreateChannelIngressMonitorOptions<TRaw, TBody, TStoredPayload, TMetadata> = {
+export type CreateChannelIngressMonitorOptions<TRaw, TBody, TStoredPayload, TMetadata> = {
   queue:
     | ChannelIngressQueue<TStoredPayload, TMetadata>
     | (() => ChannelIngressQueue<TStoredPayload, TMetadata>);
@@ -769,64 +769,5 @@ export function createChannelIngressMonitor<TRaw, TBody, TStoredPayload, TMetada
     waitForPumpIdle,
     isRunning: () => running,
     isStopped: () => stopped,
-  };
-}
-
-type StandardRawEventPayload = { version: 1; rawEvent: string };
-type StandardRawEventAdmission<TInspection> =
-  | { kind: "invalid"; message: string }
-  | { kind: "durable" | (null extends TInspection ? "ignored" : never) };
-type StandardRawEventIngressOptions<TRaw, TMetadata, TInspection> = Omit<
-  CreateChannelIngressMonitorOptions<TRaw, string, StandardRawEventPayload, TMetadata>,
-  "admissionMode" | "drain" | "inspect" | "payload" | "pollIntervalMs" | "retention"
-> & {
-  inspect: (raw: TRaw) => TInspection;
-  payload: Omit<
-    ChannelIngressMonitorPayloadCodec<TRaw, string, StandardRawEventPayload, TMetadata>,
-    "storage" | "version"
-  >;
-  pollIntervalMs?: number;
-  drain?: Omit<ChannelIngressMonitorDrainOptions<StandardRawEventPayload, TMetadata>, "startLimit">;
-  classifyAdmissionError: (error: unknown) => string | undefined;
-};
-
-/** Version-1 raw events, 500 ms polling, eight deliveries, and standard retention. */
-export function createStandardRawEventIngressMonitor<
-  TRaw,
-  TMetadata,
-  TInspection extends ChannelIngressMonitorFacts | null,
->(options: StandardRawEventIngressOptions<TRaw, TMetadata, TInspection>) {
-  const { classifyAdmissionError, createStoppedError, drain, payload, pollIntervalMs, ...base } =
-    options;
-  const stoppedError =
-    createStoppedError ?? (() => new Error("Channel ingress monitor is stopped."));
-  const monitor = createChannelIngressMonitor({
-    ...base,
-    inspect: options.inspect,
-    payload: { ...payload, storage: "raw-event", version: 1 },
-    pollIntervalMs: pollIntervalMs ?? 500,
-    retention: "standard",
-    drain: { ...drain, startLimit: 8 },
-    admissionMode: "while-running",
-    createStoppedError: stoppedError,
-  });
-  return {
-    receive: async (raw: TRaw): Promise<StandardRawEventAdmission<TInspection>> => {
-      if (!monitor.isRunning()) throw stoppedError();
-      let facts: TInspection;
-      try {
-        facts = options.inspect(raw);
-      } catch (error) {
-        const message = classifyAdmissionError(error);
-        if (message === undefined) throw error;
-        return { kind: "invalid", message };
-      }
-      if (!facts) return { kind: "ignored" } as StandardRawEventAdmission<TInspection>;
-      await monitor.admit(raw, { facts });
-      return { kind: "durable" };
-    },
-    start: monitor.start,
-    stop: monitor.stop,
-    waitForIdle: monitor.waitForIdle,
   };
 }
